Introudced Enum for Handling the Code for the Create Job API Test
Reduced Magic Number Issue in the test script

+++ src/test/java/com/api/utils/DateTimeUtil.java
@@ -11,6 +11,6 @@ public class DateTimeUtil {
 	}
 	
 	public static String getTimeWithDaysAgo(int days) {
-	    return Instant.now().minus((long) days).toString();
+	    return Instant.now().minus(days).toString();
 	}
 }
